+++ src/pages/OAuthPages/OAuthLogout.tsx
@@ -1,0 +1,6 @@
+import { FC } from 'react';
+type OAuthLogoutProps = {};
+
+export const OAuthLogout: FC<OAuthLogoutProps> = ({}) => {
+  return <div></div>;
+};
