+++ src/pages/index.ts
@@ -3,6 +3,7 @@ export * from './Introduction/Introduction';
 export * from './MainPage';
 export * from './DetailPage';
 export * from './NotFoundPage/NotFoundPage';
-export * from './OAuthPage/OAuthPage';
 export * from './TestPage/TestPage';
-export * from './OAuthPage/OAuthCallbackPage';
+export * from './OAuthPages/OAuthPage';
+export * from './OAuthPages/OAuthCallbackPage';
+export * from './OAuthPages/OAuthLogout';
