+++ src/router/Router.tsx
@@ -8,6 +8,7 @@ import {
   NotFoundPage,
   TestPage,
   OAuthCallbackPage,
+  OAuthLogout,
 } from '../pages';
 import { Header } from '../shared/Header';
 import { EditPost } from '../pages/EditPost';
@@ -29,6 +30,7 @@ export const RouterProvider = () => {
           </Route>
           <Route path="/oauth" element={<OAuthPage />} />
           <Route path="/oauth/callback" element={<OAuthCallbackPage />} />
+          <Route path="/oauth/logout" element={<OAuthLogout />} />
         </Routes>
       </ModalProvider>
     </BrowserRouter>
